requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
mysql-connector-python
